# contracts/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Condux Control Plane API",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8099/"
    }
  ],
  "paths": {
...
    "/api/orgs/{orgId}/billing": {
      "get": {
        "tags": [
          "Billing"
        ],
        "operationId": "billingStatus",
        "parameters": [
          {
            "name": "orgId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingStatusResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/orgs/{orgId}/billing/checkout": {
      "post": {
        "tags": [
          "Billing"
        ],
        "operationId": "billingCheckout",
        "parameters": [
          {
            "name": "orgId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/orgs/{orgId}/billing/portal": {
      "post": {
        "tags": [
          "Billing"
        ],
        "operationId": "billingPortal",
        "parameters": [
          {
            "name": "orgId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/stripe/webhook": {
      "post": {
        "tags": [
          "Billing"
        ],
        "operationId": "stripeWebhook",
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BillingStatusResponse": {
        "required": [
          "tier",
          "billingEnabled",
          "purchasableTiers",
          "hasSubscription"
        ],
        "type": "object",
        "properties": {
          "tier": {
            "type": "string"
          },
          "billingEnabled": {
            "type": "boolean"
          },
          "purchasableTiers": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "hasSubscription": {
            "type": "boolean"
          }
        }
      },
      "CheckoutRequest": {
        "required": [
          "tier"
        ],
        "type": "object",
        "properties": {
          "tier": {
            "type": "string"
          }
        }
      },
      "CheckoutResponse": {
        "required": [
          "url"
        ],
        "type": "object",
        "properties": {
          "url": {
            "type": "string"
          }
        }
      },
...
      "ErrorResponse": {
        "required": [
          "error"
        ],
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}